Moved nested Compass class in dedicated class, Added custom CompassView
